move media list mode enum
