+++ app/src/ereferrals/java/org/eyeseetea/malariacare/domain/boundary/repositories/ISettingsRepository.java
@@ -1,9 +1,10 @@
 package org.eyeseetea.malariacare.domain.boundary.repositories;
-
-import org.eyeseetea.malariacare.domain.entity.MediaListMode;
 import org.eyeseetea.malariacare.domain.entity.Settings;
 
 public interface ISettingsRepository {
+
+    enum MediaListMode { GRID, LIST }
+
     Settings getSettings();
 
     void saveSettings(Settings settings);
