--- app/src/main/java/org/eyeseetea/malariacare/domain/entity/MediaListMode.java
@@ -1,3 +1,0 @@
-package org.eyeseetea.malariacare.domain.entity;
-
-public enum MediaListMode { GRID, LIST }
